@@ -3,27 +3,44 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>org.example</groupId>
   <artifactId>bystro_i_tochka_maven</artifactId>
-  <packaging>war</packaging>
-  <version>1.0-SNAPSHOT</version>
+  <version>0.1.0-SNAPSHOT</version>
   <name>bystro_i_tochka_maven Maven Webapp</name>
-  <url>http://maven.apache.org</url>
+
+    <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+        <java.version>17</java.version>
+        <lombok.version>1.18.26</lombok.version>
+
+        <maven.clean.version>3.2.0</maven.clean.version>
+        <maven.compilier.version>3.11.0</maven.compilier.version>
+    </properties>
+
   <dependencies>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>3.8.1</version>
-      <scope>test</scope>
-    </dependency>
+      <dependency>
+          <groupId>org.projectlombok</groupId>
+          <artifactId>lombok</artifactId>
+          <version>${lombok.version}</version>
+          <scope>provided</scope>
+      </dependency>
   </dependencies>
+
   <build>
     <finalName>bystro_i_tochka_maven</finalName>
       <plugins>
           <plugin>
               <groupId>org.apache.maven.plugins</groupId>
+              <artifactId>maven-clean-plugin</artifactId>
+              <version>${maven.clean.version}</version>
+          </plugin>
+
+          <plugin>
+              <groupId>org.apache.maven.plugins</groupId>
               <artifactId>maven-compiler-plugin</artifactId>
+              <version>${maven.compilier.version}</version>
               <configuration>
-                  <source>8</source>
-                  <target>8</target>
+                  <source>${java.version}</source>
+                  <target>${java.version}</target>
               </configuration>
           </plugin>
       </plugins>
