@@ -7,14 +7,56 @@ public class Order {
 
     private User user;
     private int id;
+    private int addressID;
 
     private Date date;
 
+    public User getUser() {
+        return user;
+    }
+
+    public void setUser(User user) {
+        this.user = user;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public int getAddressID() {
+        return addressID;
+    }
+
+    public void setAddressID(int addressID) {
+        this.addressID = addressID;
+    }
+
+    public Date getDate() {
+        return date;
+    }
+
+    public void setDate(Date date) {
+        this.date = date;
+    }
+
+    public List<Food> getOrder() {
+        return order;
+    }
+
+    public void setOrder(List<Food> order) {
+        this.order = order;
+    }
+
     private List<Food> order;
 
-    public Order(User user, int id, Date date, List<Food> order) {
+    public Order(User user, int id, int addressID, Date date, List<Food> order) {
         this.user = user;
         this.id = id;
+        this.addressID = addressID;
         this.date = date;
         this.order = order;
     }
